requirements.txt
aiohttp
alembic
beautifulsoup4
colorlog
fastapi
httpx
jinja2
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
langgraph
loguru
numpy
pandas
psycopg2-binary
pydantic
pydantic-settings
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
redis
requests
scrapy
sentence-transformers
sqlalchemy
uvicorn
black
flake8
pytest
pytest-asyncio
ruff